Fix selection symbol (+) lost after a few seconds
This is fix on Linux. Someone has to test it on BSD and fix IF there's
similar issue with the kevent based mechanism.

We need to mask and check if the intended events are received because
some events like IN_UNMOUNT, IN_Q_OVERFLOW or IN_IGNORED may arrive
without knowledge.