Always detach opener
This may affect issue #227 but the defunct processes do not use any memory and exit when nnn exits. I guess that happens because of xdg-open using an intermediate script.

However, if we do not detach xdg-open, the child process is killed when the terminal is closed which is a severe problem.