Use KEY_RESIZE when prompting user input
The problem was that a window wouldn't resize while prompting the user
for input. These changes allows the input methods in nnn to resize
properly if KEY_RESIZE is defined.

A more portable solution would be hooking the WINCH signal and update
the xlines value from there along with some resized flag.

In some cases the full window isn't redrawn until the user has finished
the input. This is because in some functions I wasn't sure the current
"path" was available.